% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mbar.R
\name{mbar}
\alias{mbar}
\title{Bar graph for multiple response questions}
\usage{
mbar(.df, .cap = NULL)
}
\arguments{
\item{.df}{A dataframe, piped from \code{multiple_response_data_prep}}

\item{.cap}{Text for a caption (optional)}
}
\value{
A bar graph
}
\description{
\code{mbar} Takes a table prepared by \code{multiple_response_data_prep},
and outputs a bar graph.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
df \%>\%
multiple_response_data_prep(dd, Q25, .full = F) \%>\%
mbar(.cap="2020 Poll, n=")


}
