% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_response_data_prep.R
\name{single_response_data_prep}
\alias{single_response_data_prep}
\title{Makes Graph of Single response questions}
\usage{
single_response_data_prep(.df, .dd, .x)
}
\arguments{
\item{.df}{An imported dataframe}

\item{.dd}{A data dictionary}

\item{.x}{A question number}
}
\value{
A tibble
}
\description{
\code{single_response_respondent_level} returns a bar graph of a given question.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
df \%>\% single_response_data_prep(dd, Q8)

}
