% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_response_respondent_level_ord.R
\name{single_response_respondent_level_ord}
\alias{single_response_respondent_level_ord}
\title{Makes Graph of Single response, ordinal questions}
\usage{
single_response_respondent_level_ord(.df, .dd, .x)
}
\arguments{
\item{.df}{An imported dataframe}

\item{.dd}{A data dictionary}

\item{.x}{The question id number of an ordinal question}
}
\value{
A bar chart
}
\description{
\code{single_response_respondent_level_ord} returns a bar graph of a given question.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
df \%>\% single_response_respondent_level_ord(dd, Q11)

}
