% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_agr_plot.R
\name{ted_agr_plot}
\alias{ted_agr_plot}
\title{Creates graph for multiple item agreement questions}
\usage{
ted_agr_plot(.df, .cap = NULL)
}
\arguments{
\item{.df}{A dataframe, piped from \code{ted_agr_prep}}

\item{.cap}{Text for a caption (optional)}
}
\value{
A graph
}
\description{
\code{ted_agr_plot} Given a tibble of a given multiple level agreement question,
    creates graph.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
ted_agr_prep(df, dd, Q27, c("Agree", "Neither agree nor disagree", "Disagree")) \%>\%
ted_agr_plot()

}
