% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_agr_plot.R, R/ted_agr_prep.R
\name{ted_agr_prep}
\alias{ted_agr_prep}
\title{Creates graph for multiple item agreement questions}
\usage{
ted_agr_prep(.df, .dd, .x, .l)

ted_agr_prep(.df, .dd, .x, .l)
}
\arguments{
\item{.df}{A dataframe, piped from \code{ted_agr_prep}}

\item{.dd}{A data dictionary}

\item{.x}{A question number}

\item{.l}{The three level names}

\item{.cap}{Text for a caption (optional)}

\item{.df}{An imported dataframe}
}
\value{
A graph

A tibble
}
\description{
\code{ted_agr_plot} Given a tibble of a given multiple level agreement question,
    creates graph.

\code{ted_agr_prep} returns a tibble of a given multiple level agreement question,
    prepared for tables or graphs.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora

This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
ted_agr_prep(df, dd, Q27, c("Agree", "Neither agree nor disagree", "Disagree")) \%>\%
ted_agr_plot()

ts_agr_prep(df, dd, Q13, c("Agree", "Neither agree nor disagree", "Disagree"))

}
