% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_con_plot.R
\name{ted_con_plot}
\alias{ted_con_plot}
\title{Creates graph for multiple item likelihood questions}
\usage{
ted_con_plot(.df, .cap = NULL)
}
\arguments{
\item{.df}{A dataframe, piped from \code{\link{ted_con_prep}}}

\item{.cap}{Text for a caption (optional)}
}
\value{
A graph
}
\description{
\code{ted_con_plot} Given a tibble of a given multiple level likelihood question,
    creates graph.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
ted_con_prep(df, dd, Q29, c("Likely", "Neither likely nor unlikely", "Unlikely")) \%>\%
ted_con_plot()

}
