% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_con_prep.R
\name{ted_con_prep}
\alias{ted_con_prep}
\title{Prepares data for multiple item likelihood questions}
\usage{
ted_con_prep(.df, .dd, .x, .l)
}
\arguments{
\item{.df}{An imported dataframe}

\item{.dd}{A data dictionary}

\item{.x}{A question number}

\item{.l}{The three level names}
}
\value{
A tibble
}
\description{
\code{ted_con_prep} returns a tibble of a given multiple level confidence question,
    prepared for tables or graphs.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
ted_con_prep(df, dd, Q29, c("Likely", "Neither likely nor unlikely", "Unlikely"))

}
