% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_sr_data_prep.R
\name{ted_sr_data_prep}
\alias{ted_sr_data_prep}
\title{Makes Graph of Single response questions}
\usage{
ted_sr_data_prep(.df, .dd, .x)
}
\arguments{
\item{.df}{An imported dataframe}

\item{.dd}{A data dictionary}

\item{.x}{A question number}
}
\value{
A tibble
}
\description{
\code{ted_sr_data_prep} returns a bar graph of a given question.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
df \%>\% ted_sr_data_prep(dd, Q8)

}
