% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_yn_plot.R
\name{ted_yn_plot}
\alias{ted_yn_plot}
\title{Creates graph for multiple yes no type questions}
\usage{
ted_yn_plot(.df, .cap = NULL)
}
\arguments{
\item{.df}{A dataframe, piped from \code{\link{ted_yn_prep}}}

\item{.cap}{Text for a caption (optional)}
}
\value{
A graph
}
\description{
\code{ted_yn_plot} Given a tibble of a given multiple level yes/no type question,
    creates graph.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
ted_yn_prep(df, dd, Q17, .l=c("Strength","Weakness")) \%>\%
ted_yn_plot()

}
