% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_yn_prep.R
\name{ted_yn_prep}
\alias{ted_yn_prep}
\title{Prepares data for multiple two item questions}
\usage{
ted_yn_prep(.df, .dd, .x, .l)
}
\arguments{
\item{.df}{An imported dataframe}

\item{.dd}{A data dictionary}

\item{.x}{A question number}

\item{.l}{The two level names}
}
\value{
A tibble
}
\description{
\code{ted_yn_prep} returns a tibble of a given multiple item yes/no like question,
    prepared for tables or graphs.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
ted_yn_prep(df, dd, Q17, .l=c("Strength","Weakness"))

}
