% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ts_agreement_plot.R
\name{ts_agreement_plot}
\alias{ts_agreement_plot}
\title{Creates graph for multiple item agreement questions}
\usage{
ts_agreement_plot(.df, .cap = NULL)
}
\arguments{
\item{.df}{A dataframe, piped from \code{ts_agreement_prep}}

\item{.cap}{Text for a caption (optional)}
}
\value{
A bar graph

A graph
}
\description{
\code{ts_agreement_plot} Given a tibble of a given multiple level agreement question,
    creates graph.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
ts_agreement_prep(df, dd, Q27, c("Agree", "Neither agree nor disagree", "Disagree")) \%>\%
ts_agreement_plot()

}
