% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ted_agr_prep.R
\name{ted_agr_prep}
\alias{ted_agr_prep}
\title{Prepares data for multiple item agreement questions}
\usage{
ted_agr_prep(.df, .dd, .x, .l)
}
\arguments{
\item{.df}{An imported dataframe}

\item{.dd}{A data dictionary}

\item{.x}{A question number}

\item{.l}{The three level names}
}
\value{
A tibble
}
\description{
\code{ted_agr_prep} returns a tibble of a given multiple level agreement question,
    prepared for tables or graphs.
}
\details{
This is an NACD function: Should you experience difficulties, see Ted Sikora
}
\examples{
ts_agr_prep(df, dd, Q13, c("Agree", "Neither agree nor disagree", "Disagree"))

}
